rewrites elementor widget to work alongside gutenberg block
