+++ webpack.config.js
@@ -1,5 +1,6 @@
 const defaultConfig = require("@wordpress/scripts/config/webpack.config.js");
 const TerserPlugin = require("terser-webpack-plugin");
+const path = require("path");
 
 module.exports = {
   ...defaultConfig,
@@ -8,5 +9,9 @@ module.exports = {
       minimize: true,
       minimizer: [new TerserPlugin()],
     },
+    entry: {
+      ...defaultConfig.entry(),
+      elementorWidget: path.join(path.resolve(), "src/elementorWidget.js")
+    }
   },
 };
